removed should.js, moved tests to subfolders, added mocha.opts
